requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
annotated-types==0.5.0
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backoff==2.2.1
bandit==1.7.5
black==23.3.0
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.26
cleo==2.0.1
click==8.1.3
clickhouse-connect==0.6.4
colorama==0.4.6
coloredlogs==15.0.1
crashtest==0.4.1
dataclasses-json==0.5.8
distlib==0.3.6
docutils==0.20.1
duckdb==0.8.1
dulwich==0.21.5
fastapi==0.99.1
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
html5lib==1.1
httptools==0.5.0
humanfriendly==10.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
langchain==0.0.201
langchainplus-sdk==0.0.10
lockfile==0.12.2
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
pbr==5.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.5.3
pluggy==1.0.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
posthog==3.0.1
protobuf==4.23.3
ptyprocess==0.7.0
pulsar-client==3.2.0
pycodestyle==2.10.0
pydantic==1.10.10
pydantic_core==2.0.1
pyflakes==3.0.1
Pygments==2.15.1
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-env-loader==0.0.2
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0
rapidfuzz==2.15.1
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.16
starlette==0.27.0
stevedore==5.1.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tomlkit==0.11.8
tqdm==4.65.0
trove-classifiers==2023.5.24
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
virtualenv==20.23.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0